Added BaseDataLoader to base package * Pulled up batch iteration functions from GaussianDataLoader to BaseDataLoader * Reference path changes in dataloader package
