+++ src/dataloaders/multi_normal.py
@@ -1,50 +1,6 @@
 import numpy as np
-
-from dataloaders.base import BaseDataLoader
+from base.dataloader import BaseDataLoader
 from utils import np_utils
-
-
-class GaussianDataLoader(BaseDataLoader):
-
-    def __init__(self, input_size=1, latent_size=1, train_batch_size=32, test_batch_size=32):
-        super(GaussianDataLoader, self).__init__(input_size, latent_size)
-
-        train_data, test_data = self.get_data()
-
-        np.random.shuffle(train_data)
-        np.random.shuffle(test_data)
-
-        self.batch_size = {
-            'train': train_batch_size,
-            'test': test_batch_size
-        }
-
-        self.data = {
-            'train': train_data,
-            'test': test_data
-        }
-
-        self.n_batches = {
-            split: self.data[split].shape[0] // self.batch_size[split]
-            for split in ['train', 'test']
-        }
-
-        self.batch_index = {
-            'train': 0,
-            'test': 0
-        }
-
-    def next_batch(self, split):
-        start = self.batch_index[split] * self.batch_size[split]
-        end = start + self.batch_size[split]
-        self.batch_index[split] = (self.batch_size[split] + 1) % self.n_batches[split]
-
-        if split == 'train' and self.batch_index[split] == 0:
-            np.random.shuffle(self.data[split])
-
-        data = self.data[split][start: end]
-
-        return data
 
 
 class TwoGaussiansDataLoader(BaseDataLoader):
@@ -105,8 +61,8 @@ class FourGaussiansDataLoader(BaseDataLoader):
 
 
 class FourSymGaussiansDataLoader(BaseDataLoader):
-    def __init__(self, input_size=1, latent_size=1):
-        super(FourSymGaussiansDataLoader, self).__init__(input_size, latent_size)
+    def __init__(self, input_size=1, latent_size=1, train_batch_size=32, test_batch_size=32):
+        super(FourSymGaussiansDataLoader, self).__init__(input_size, latent_size, train_batch_size, test_batch_size)
 
     def get_data(self, train_ratio=0.6):
         num_samples = 20000
